# swagger.yaml
openapi: 3.0.3
info:
  title: ProcSentinel API
  description: |
    REST API for managing blocked applications and server status in the ProcSentinel process monitoring system.
    
    ## Features
    - Manage blocked applications list
    - Enable/disable server functionality
    - Persistent SQLite storage with in-memory caching
    - Thread-safe operations
    
    ## Storage
    The server uses SQLite database (`./server/procsentinel.db`) with in-memory caching for optimal performance.
    When the server is disabled, the applications list returns empty but data persists in the database.
  version: 1.0.0
  contact:
    name: ProcSentinel
    url: https://github.com/procsentinel
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check endpoint
      description: Simple health check to verify server is running
      operationId: getHealth
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
              example:
                status: ok

  /applications:
    get:
      tags:
        - Applications
      summary: Get list of blocked applications
      description: |
        Returns the list of all blocked applications. 
        If the server is disabled, returns an empty list even if applications exist in the database.
      operationId: getApplications
      responses:
        '200':
          description: List of blocked applications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationsResponse'
              examples:
                enabled:
                  summary: Server enabled with applications
                  value:
                    applications: ["firefox", "chrome", "opera"]
                disabled:
                  summary: Server disabled (empty list)
                  value:
                    applications: []
                empty:
                  summary: No blocked applications
                  value:
                    applications: []

    post:
      tags:
        - Applications
      summary: Add new blocked application
      description: |
        Adds a new application to the blocked list. The application name must be unique and non-empty.
        Data is stored in both cache and SQLite database.
      operationId: addApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Application'
            examples:
              firefox:
                summary: Block Firefox
                value:
                  name: firefox
              chrome:
                summary: Block Chrome
                value:
                  name: chrome
      responses:
        '201':
          description: Application added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Application 'firefox' added to blocked list"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                empty_name:
                  summary: Empty application name
                  value:
                    error: "Application name cannot be empty"
                invalid_json:
                  summary: Invalid JSON
                  value:
                    error: "Invalid JSON"
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Method not allowed"
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Failed to save to database: ..."

  /applications/{name}:
    delete:
      tags:
        - Applications
      summary: Remove blocked application
      description: |
        Removes an application from the blocked list by name. 
        The application must exist in the list to be removed.
        Data is removed from both cache and SQLite database.
      operationId: removeApplication
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the application to remove
          schema:
            type: string
            example: firefox
      responses:
        '200':
          description: Application removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Application 'firefox' removed from blocked list"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Application name is required in URL path"
        '404':
          description: Application not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Application 'nonexistent' not found in blocked list"
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Method not allowed"
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Failed to remove from database: ..."

  /status:
    get:
      tags:
        - Status
      summary: Get server status
      description: |
        Returns the current server status (enabled/disabled).
        When disabled, the server continues to run but returns empty application lists.
      operationId: getStatus
      responses:
        '200':
          description: Current server status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                enabled:
                  summary: Server is enabled
                  value:
                    enabled: true
                disabled:
                  summary: Server is disabled
                  value:
                    enabled: false
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Method not allowed"

    put:
      tags:
        - Status
      summary: Update server status
      description: |
        Updates the server status (enable/disable). 
        When disabled, GET /applications returns empty list but data persists in database.
        Status is stored in both cache and SQLite database.
      operationId: updateStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusResponse'
            examples:
              enable:
                summary: Enable server
                value:
                  enabled: true
              disable:
                summary: Disable server
                value:
                  enabled: false
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              examples:
                enabled:
                  summary: Server enabled
                  value:
                    message: "Server enabled"
                disabled:
                  summary: Server disabled
                  value:
                    message: "Server disabled"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Invalid JSON"
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Method not allowed"
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Failed to save status to database: ..."

  /reset:
    delete:
      tags:
        - Applications
      summary: Reset blocked applications list
      description: |
        Removes all applications from the blocked list. This clears both the in-memory cache 
        and removes all records from the SQLite database. The operation is atomic and 
        thread-safe.
      operationId: resetApplications
      responses:
        '200':
          description: All applications removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              examples:
                reset_with_apps:
                  summary: Reset with applications present
                  value:
                    message: "Reset complete: removed 3 applications from blocked list"
                reset_empty:
                  summary: Reset with no applications
                  value:
                    message: "Reset complete: removed 0 applications from blocked list"
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Method not allowed"
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Failed to reset applications in database: ..."

components:
  schemas:
    Application:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the application to block
          example: firefox
          minLength: 1
          maxLength: 255
      example:
        name: firefox

    ApplicationsResponse:
      type: object
      required:
        - applications
      properties:
        applications:
          type: array
          items:
            type: string
          description: List of blocked application names
          example: ["firefox", "chrome", "opera"]
      example:
        applications: ["firefox", "chrome", "opera"]

    StatusResponse:
      type: object
      required:
        - enabled
      properties:
        enabled:
          type: boolean
          description: Whether the server is enabled or disabled
          example: true
      example:
        enabled: true

    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Success or informational message
          example: "Application 'firefox' added to blocked list"
      example:
        message: "Application 'firefox' added to blocked list"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message describing what went wrong
          example: "Application name cannot be empty"
      example:
        error: "Application name cannot be empty"

tags:
  - name: Health
    description: Health check operations
  - name: Applications
    description: Operations for managing blocked applications
  - name: Status
    description: Operations for managing server status

externalDocs:
  description: ProcSentinel Documentation
  url: ./WARP.md